requirements.txt
matplotlib
numpy
opencv-python
pandas
pillow
PySide6
pytest
PyYAML
reportlab
sqlalchemy
tqdm
ultralytics
